Optimized search/term selection for badges
